requirements.txt
aiodns==4.0.0
aiohttp==3.13.3
aiomysql==0.3.2
brotli==1.2.0
brotlicffi==1.2.0.0
cryptography==46.0.5
email-validator==2.3.0
fastapi-cli==0.0.20
githubkit==0.14.6
h2==4.3.0
httptools==0.7.1
nonebot-adapter-onebot==2.4.6
nonebot-plugin-alconna==0.60.4
nonebot-plugin-apscheduler==0.5.0
nonebot-plugin-htmlrender==0.6.7
nonebot-plugin-orm==0.8.3
nonebot-plugin-sentry==2.0.0
nonebot-plugin-status==0.9.0
nonebot-plugin-uninfo==0.10.1
openai==2.26.0
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pyjwt==2.11.0
pysocks==1.7.1
python-multipart==0.0.22
tomli==2.4.0
typer-slim==0.20.0
watchfiles==1.1.1
websockets==16.0
e2b_code_interpreter==2.4.1
pymysql==1.1.2